+++ src/index.ts
@@ -6,7 +6,7 @@ import { constructConstructorProperty } from "./rules/construct-constructor-prop
 import { noConstructInInterface } from "./rules/no-construct-in-interface";
 import { noConstructStackSuffix } from "./rules/no-construct-stack-suffix";
 import { noImportPrivate } from "./rules/no-import-private";
-import { noMutablePropsInterface } from "./rules/no-mutable-props-interface";
+import { noMutablePropertyOfPropsInterface } from "./rules/no-mutable-property-of-props-interface";
 import { noMutablePublicPropertyOfConstruct } from "./rules/no-mutable-public-property-of-construct";
 import { noParentNameConstructIdMatch } from "./rules/no-parent-name-construct-id-match";
 import { noPublicClassFields } from "./rules/no-public-class-fields";
@@ -26,7 +26,7 @@ const rules = {
   "require-passing-this": requirePassingThis,
   "no-variable-construct-id": noVariableConstructId,
   "no-mutable-public-property-of-construct": noMutablePublicPropertyOfConstruct,
-  "no-mutable-props-interface": noMutablePropsInterface,
+  "no-mutable-property-of-props-interface": noMutablePropertyOfPropsInterface,
   "construct-constructor-property": constructConstructorProperty,
   "require-jsdoc": requireJSDoc,
   "require-props-default-doc": requirePropsDefaultDoc,
@@ -66,7 +66,7 @@ const recommended = createFlatConfig({
   "cdk/require-passing-this": ["error", { allowNonThisAndDisallowScope: true }],
   "cdk/no-variable-construct-id": "error",
   "cdk/no-mutable-public-property-of-construct": "warn",
-  "cdk/no-mutable-props-interface": "warn",
+  "cdk/no-mutable-property-of-props-interface": "warn",
   "cdk/construct-constructor-property": "error",
 });
 
@@ -82,7 +82,7 @@ const strict = createFlatConfig({
   "cdk/require-passing-this": "error",
   "cdk/no-variable-construct-id": "error",
   "cdk/no-mutable-public-property-of-construct": "error",
-  "cdk/no-mutable-props-interface": "error",
+  "cdk/no-mutable-property-of-props-interface": "error",
   "cdk/construct-constructor-property": "error",
   "cdk/require-jsdoc": "error",
   "cdk/require-props-default-doc": "error",
